+++ lionagi/protocols/adapters/adapter.py
@@ -1,6 +1,7 @@
 # Copyright (c) 2023 - 2024, HaiyangLi <quantocean.li at gmail dot com>
 #
 # SPDX-License-Identifier: Apache-2.0
+
 import logging
 from typing import Any, Protocol, TypeVar, runtime_checkable
