+++ lionagi/service/providers/openai_/chat_completions.py
@@ -2,9 +2,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from lionagi.service.api.endpoints.chat_completion import (
-    ChatCompletionEndPoint,
-)
+from lionagi.service.endpoints.chat_completion import ChatCompletionEndPoint
 
 CHAT_COMPLETION_CONFIG = {
     "provider": "openai",
